Refactor and tidy.
